requirements.txt
opencv-python
tensorflow
Pillow
tqdm
scikit-learn
numpy
matplotlib
keras-cv-attention-models
